Rename ops/ and types/agent_types.go
